requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
greenlet==3.0.3

# HTTP client
aiohttp==3.9.3

# Settings management
pydantic-settings==2.1.0

# Scheduling
apscheduler==3.10.4

# Date/time handling
python-dateutil==2.8.2
